refactor(backend): colocate presentation tests
